requirements.txt
fastapi
uvicorn
python-multipart
pillow
sqlalchemy
psycopg2-binary
httpx
python-jose[cryptography]
passlib[bcrypt]
python-dotenv